#503 New startup list items
Test new timer for TSAT, will show time to TSAT in minutes or time past TSAT in minutes.

New STS field, for network status. It will allow us to send REA, PRIO or show SUSP for aircrafts. REA will attempt to find a better CTOT and only give it if it is better. PRIO will set the aircraft first of all with CTOT. SUSP is set automatically when flightplan is suspended due to TOBT, TSAT or other cases from the server side.